Split key-info out of key command.
The key-info command reads an exported ascii-armored private key from
file and displays the info that's useful in assertions (specifically an
account-key assertion). This is really useful when:

* you already have a key, e.g. from gnupg
* you generated a key but didn't take note of the useful bits ;)

The key command can still be used for convenience but it's probably
better to use a key generated by something like gnupg.